backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
pydantic
langchain
langchain-community
langchain-google-genai
google-ai-generativelanguage==0.4.0
